+++ src/client/commands/ps.rs
@@ -1,6 +1,6 @@
 use crate::client::cli::ClientConfig;
 use crate::client::request;
-use crate::daemon::api::ApiModuleRunStatus;
+use crate::daemon::api::{ApiModuleRunStatus, ApiProbeStatus};
 use anyhow::Result;
 use chrono::Local;
 use std::convert::TryFrom;
@@ -13,7 +13,7 @@ pub fn list_modules_cmd(cfg: &ClientConfig) -> Result<()> {
     let module_status = request::list_modules(&cfg.daemon_url)?;
     let mut tw = TabWriter::new(io::stdout()).minwidth(8);
 
-    writeln!(&mut tw, "pid\tname\tstatus\tsince")?;
+    writeln!(&mut tw, "pid\tname\tliveness\tstatus\tsince")?;
     module_status.status.iter().try_for_each(|mod_status| {
         let formatted_status = match mod_status.status {
             ApiModuleRunStatus::RUNNING => "running",
@@ -21,6 +21,15 @@ pub fn list_modules_cmd(cfg: &ClientConfig) -> Result<()> {
             ApiModuleRunStatus::WAITING => "waiting",
             ApiModuleRunStatus::EXITED => "exited",
         };
+        let formatted_liveness_status = match mod_status.liveness_status {
+            Some(ApiProbeStatus::Pending) => "pending",
+            Some(ApiProbeStatus::Successful) => "healthy",
+            Some(ApiProbeStatus::Failing) => "failing",
+            Some(ApiProbeStatus::Error) => "erroring",
+            Some(_) => "none",
+            None => "none",
+        };
+
         let time_formatter = timeago::Formatter::new();
         let now = u64::try_from(Local::now().timestamp()).unwrap();
         let dur = Duration::new(now - mod_status.time_since_status, 0);
@@ -33,8 +42,12 @@ pub fn list_modules_cmd(cfg: &ClientConfig) -> Result<()> {
 
         writeln!(
             &mut tw,
-            "{}\t{}\t{}\t{}",
-            mod_status.pid, mod_status.name, formatted_status, formatted_time,
+            "{}\t{}\t{}\t{}\t{}",
+            mod_status.pid,
+            mod_status.name,
+            formatted_liveness_status,
+            formatted_status,
+            formatted_time,
         )
     })?;
     tw.flush()?;