+++ src/daemon/monitor/runtime.rs
@@ -1,10 +1,10 @@
 use super::commands::*;
-use super::poll::{channel_rx, poll_tickr};
+use super::poll::{channel_rx, liveness_poll_tickr, readiness_poll_tickr};
 use super::state::{MonitorState, MonitorStatus};
 use anyhow::Result;
 use log::info;
-use std::sync::Arc;
 use std::thread;
+use std::{collections::HashMap, sync::Arc};
 use tokio::runtime::{self, Handle, Runtime};
 use tokio::sync::mpsc;
 
@@ -30,10 +30,29 @@ impl MonitorHandle {
         }
     }
 
-    pub fn new_monitor(&self, name: String, monitor: Monitor) {
+    pub fn new_monitor(
+        &self,
+        name: String,
+        monitor: Monitor,
+        monitor_type: MonitorType,
+    ) {
         let tx = self.producer.clone();
         self.runtime_handle.spawn(async move {
-            let cmd = MonitorCommand::NewMonitor { key: name, monitor };
+            let cmd = MonitorCommand::NewMonitor {
+                key: name,
+                monitor,
+                monitor_type,
+            };
+            tx.send(cmd)
+                .await
+                .expect("Failed to transmit to monitor runtime");
+        });
+    }
+
+    pub fn remove_monitor(&self, key: String, monitor_type: MonitorType) {
+        let tx = self.producer.clone();
+        self.runtime_handle.spawn(async move {
+            let cmd = MonitorCommand::RemoveMonitor { key, monitor_type };
             tx.send(cmd)
                 .await
                 .expect("Failed to transmit to monitor runtime");
@@ -42,6 +61,20 @@ impl MonitorHandle {
 
     pub fn monitor_status(&self, monitor_name: &str) -> Option<MonitorStatus> {
         self.monitor_state.monitor_status(monitor_name)
+    }
+
+    pub fn monitor_statuses(&self) -> HashMap<String, MonitorStatus> {
+        self.monitor_state.monitor_statuses()
+    }
+}
+
+impl Clone for MonitorHandle {
+    fn clone(&self) -> Self {
+        MonitorHandle {
+            runtime_handle: self.runtime_handle.clone(),
+            producer: self.producer.clone(),
+            monitor_state: self.monitor_state.clone(),
+        }
     }
 }
 
@@ -53,9 +86,21 @@ fn setup_runtime() -> Result<Runtime, std::io::Error> {
         .build()
 }
 
+pub fn monitor_key(name: &str, monitor_type: &MonitorType) -> String {
+    match monitor_type {
+        MonitorType::Liveness => {
+            format!("{}-{}-liveness", name, uuid::Uuid::new_v4())
+        }
+        MonitorType::Readiness => {
+            format!("{}-{}-readiness", name, uuid::Uuid::new_v4())
+        }
+    }
+}
+
 pub fn spawn_runtime(monitor_state: Arc<MonitorState>) -> MonitorHandle {
     let (tx, rx) = mpsc::channel::<MonitorCommand>(32);
-    let tx_ = tx.clone();
+    let tx_readiness = tx.clone();
+    let tx_liveness = tx.clone();
     let (handle_tx, handle_rx) = std::sync::mpsc::channel();
     let mst = Arc::clone(&monitor_state);
 
@@ -68,8 +113,10 @@ pub fn spawn_runtime(monitor_state: Arc<MonitorState>) -> MonitorHandle {
             .send(runtime.handle().clone())
             .expect("Unable to give runtime handle to main thread");
 
-        // Spawn the ticking thread for scanning of monitors
-        runtime.spawn(async move { poll_tickr(tx_).await });
+        // Spawn the ticking thread for scanning of readiness monitors
+        runtime.spawn(async move { readiness_poll_tickr(tx_readiness).await });
+        // Spawn the ticking thread for scanning of liveness monitors
+        runtime.spawn(async move { liveness_poll_tickr(tx_liveness).await });
 
         // Continue running until notified to shutdown
         runtime.block_on(async { channel_rx(rx, mst).await });